:truck: Split up ScnearioCard file.
